+++ HISTORY.rst
@@ -116,6 +116,7 @@ Unreleased
 - #940 Set/change font color when working with Hyperlinks is impossible
   (add ``Font.use_theme_hyperlink_color`` flag and
   ``ThemePart.theme.hlink_color`` / ``folHlink_color`` accessors)
+- #201 slide numbers / date / footer via p:hf and a:fld
 
 1.0.2 (2024-08-07)
 ++++++++++++++++++